Sanity checking

+++ pantomath.py
@@ -72,10 +72,10 @@ class MainFrame(wx.Frame):
     def ImportFromInbox(self, filepath):
         if self.InboxHasSettled():
             wx.LogDebug(f'Importing {filepath}')
-            self.library.import_file(filepath)
-            if self.config.Read("/Import/RemoveSource", ''):
-                wx.LogDebug(f'Removing {filepath}')
-                os.remove(filepath)
+            if self.library.import_file(filepath):
+                if self.config.Read("/Import/RemoveSource", ''):
+                    wx.LogDebug(f'Removing {filepath}')
+                    os.remove(filepath)
         else:
             wx.LogDebug('Waiting for Inbox to settle')
             wx.CallLater(1000, self.ImportFromInbox, filepath)
